+++ FlightGear.h
@@ -10,11 +10,11 @@
 // between the client and FlightGear very easy
 typedef struct {
     int magnetos = 0;
-    double mixture = 0.0;
-    double throttle = 0.0;
-    double aileron = 0.0;
-    double elevator = 0.0;
-    double rudder = 0.0;
+    float mixture = 0.0;
+    float throttle = 0.0;
+    float aileron = 0.0;
+    float elevator = 0.0;
+    float rudder = 0.0;
 } FGFS_Data;
 
 class FlightGear {
@@ -24,11 +24,11 @@ class FlightGear {
     public:
         FlightGear(int port, char* addr = "127.0.0.1");
         ~FlightGear();
-        FlightGear* setAlieron(double aileron);
-        FlightGear* setElevator(double elevator);
-        FlightGear* setRudder(double rudder);
-        FlightGear* setThrottle(double throttle);
-        FlightGear* setMixture(double mixture);
+        FlightGear* setAlieron(float aileron);
+        FlightGear* setElevator(float elevator);
+        FlightGear* setRudder(float rudder);
+        FlightGear* setThrottle(float throttle);
+        FlightGear* setMixture(float mixture);
         FlightGear* setMagnetos(int magnetos);
         FlightGear* resetData();
         FGFS_Data getData();
